// client/src/types/UpdateHabitRequest.ts
import type { HabitDifficulty } from './HabitDifficulty'
import type { HabitFrequencyType } from './HabitFrequencyType'

export type UpdateHabitRequest = {
  name: string
  description: string | null
  category: string | null
  frequencyType: HabitFrequencyType
  targetCount: number
  difficulty: HabitDifficulty
}